More coverage code review updates.
